Update README.md
These are a complete set of files for flashing Ender 5 pro, controlled by skrmini v3.0 and tft35 e3 v3.0 display. CRtouch sample for the bed. Everything is tested and works! Just be careful with the motor step settings! At least check this carefully at the start of your printer if they match exactly! If necessary, open the files with VScode and edit as you need if necessary! Do not touch the original files! First copy the folders with the files and edit them so that you have the option to go back!

Many options and settings are enabled! The printer works great. Both display modes work, Reprap normal and Touchscreen mode. It took me two weeks to compile and test things to get here!